Manual merge of the original ticket changes into the latest commits, as there was a lot of development code to get in the way.
Changes made- > random visual player and related media objects now have react components with the controls and behavior necessary for them
Scene listener now updates the media queue when it will mount, to have the correct information flow at the right event.
Video-media preview player now has the require vimeo id module to be able to load vimeo previews -> was broken.
Embedded Vimeo player _> on MessagedReceived now checks if the data send to it is a string, as there was a situation when if a method send the message it was an object but if an event send the message it was a string.
The vimeo url now updated to use the correct options, as api 1, was stopping some events from vimeo to come through -> finished, play progress