Large Refactor - ES6, Flux lib and Store class replacement
Using the flux library, the Dispatcher has been replaced with one written by external party.

Refactored all the stores to also use the flux provided classes for integrated use with the Dispatcher

Smaller refactors, reformat comments and some code
Updated the connection cache to use a static variable for the groupNames rather than relying on the object to have this as an instance variable.