Adjustable layout
Default page for the index of the web app has been changed to load the adjustable grid component

All standalone components that are using handlers and properties passed to them by the a-grid component do not work as standalone at the moment.
Currently components that are loaded in are hardcoded in the GridStore's state, this is will probably change to facilitate the user's self layout building idea.
The layout parent component is resizable through a little re-size handler on the bottom right side of each component parent container.

3 currently not working button representing a minimize, maximize and restore buttons,
- They should change the size and positioning of the component , which would re-render the grid with the new sizing.
Minimize - should make the component as be the smallest possible size for the grid, while not changing the position
Maximize - should expand the component to the full width and height for the grid , setting to top left position 0,0
Restore - should return the component to its original size and position

Current issue with the react-layout-grid library:
When the layout property gets a change to any of the required grid layout props. (x,y,w,h), It does not trigger a re-render of the actual grid-layout.
Investigation leads to the library used within the react-grid-layout library.
The function is an npm library for the lodash isEquals  https://lodash.com/docs/4.17.4#isEqual, but when comparing the current array and the new incoming one, which have different size and position data - the function says that the arrays are the same.
This does not re-render the layout with its new sizes.

 ToDo's:
-Change the conditional statements for the layout components that were looking for location based properties
- Add the autocomplete node searcher to the graph
- use the spacing algorithm in the graph
- Fix the draggable component to remove some unnecessary function that have currently been set directly in the main jsx page of the grid.