Added further support and refactor around playing back of our own audio files.
Included some additional cleanup required during removing the player.  Since the player is no longer a single page, we cannot rely on the browser to clean up.

Added some additional utilities for better sharing of helper functions and they are easy to unit test in a standalone util file rather than with a React component.