Allow media objects and the player to clean up if the component is unmounted.
We never had this issue before as the browser would do this cleanup for us.  Now as the component can be removed from the grid, we must support this kind of unmounting.

I don't think this is perfect but it's a start and is along the right lines while keeping no-ops out of the dev console.