Allow maxFiles to reset without the component needing to be reinitialised.
Requires changes to the underlying react wrapper for the resumable-js code.
Currently these changes live in a forked version of the lib.
A pull request will be made to incorporate the changes into the lib if possible.
We may need to adjust slightly to allow the functionality we require to be toggleable.