Sound graph dashboard for Robin,
It contains a new type of UI that is created by the scene graph, that is used to send score commands to the Player, by toggling on and off the themes in a SceneGraphList.

Functions: Dynamically load and bind all themes and their first and second tag to a 2 dimensional table.

The name of the themes as the v1 of this commit should match tag1 + tag2 in order for the cell related to be able to do a score command on toggle.