Avoid incorrect set scene state in scene-listener.
The duplicate or over usage of set state of the scene has caused us to lose the queue, active and masterlist state important when changing tag matcher but not scene.

We only want to setScene on a queue when we are actually changing scene.  If we are still using the same scene, we must not recalculate the masterlist, queue and active.