Cherry pick 57e4135 Updated lodash dependecy to use new available functionality  such as uniqBy. Category creation is now moved to its own file. The Scene Graph Update action is now used to update the scene graph, whenever there are changes to to category configuration.
Current issues: the value being send back to the scene graph editor and graph does not contain the latest category config settu, while the values  are in the database and load upon refresh.
minor issue - the mapping of the cells of the table needs to be changed from this hardcoded string combination comparing.