Update to the sound graph behaviour and bugfixes:
The sound graph will now correctly build its themes and scene up and will correctly remove the correct data from thee score command when themes are toggled off.
The rootNode data is now provided to the graph so that the toggle handler can get access to the relationship data build by the parent component.
BugFix: updated the chain syntax implementations of lodash to match the the latest version's.
Sending a scoreCommand without scenes will
clear the active scene from the sceneListener forcing it to stop playback. This is reusable for future work.