Ammends to previous PR request.
Test for when the scene graph is not of the type sound have not been added due to where and how the function is used, the check happens before the generators are used.
Category config generator has been amended to give some information on the vague lines of code, bad logic has been removed before removing duplication of strings.
Graph theme generator has been amended to not have duplicate code while maintaining the option to have a scene store reference for testing