Resolved Monaco editor issue, previous history attempt was not successful.
Using sha1 and storing a short history of every version of the scene that is opened in the editor.
sha1 used for reduced errors for equality checks (maybe a non issue but reliable)
History is appended now in two places, where the editor is forced update, ie an outside update, such as adding a new media object.  The history is also appended too when the editor itself tries to save a copy, this was prone until I used clone data at the point of save.  This race condition could allow the history to have an incorrect version of the scene saved.