Ensure the theme and tag selectors are children to the aspect ratio sized player.
Without them being children of the player, the position absolute and bottom 0 leads to them being out of the player for displays that do not match the aspect ratio.

This leads us to an interesting problem to resolve.  Do we need another strategy for positioning an aspect ratio sized player within a non matching window.